requirements.txt
anthropic
gradio
jiwer
openai
python-dotenv
sarvamai
sounddevice
soundfile
